backend-v2/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib==1.7.4
bcrypt==4.0.1
pydantic
alembic